requirements_dev.txt
-r requirements_common.txt

# Tests
coverage==5.5
pylint==2.5.3
pylint-django==2.1.0

# Documentation
Sphinx==3.5.4
sphinx-rtd-theme==0.5.2
sphinx-autodoc-typehints==1.12.0
